# server/config/swagger.json
{
   "swagger": "2.0",
   "info": {
      "version": "1.0.0",
      "title": "nodejsexpressapp API"
   },
   "consumes": [
      "text/plain"
   ],
   "produces": [
      "application/json"
   ],
   "paths": {
      "/health": {
         "get": {
            "operationId": "get",
            "description": "Get health status of nodejsmicroservice",
            "responses": {
               "200": {
                  "description": "Health check response",
                  "schema": {
                     "$ref": "#/definitions/healthResponse"
                  },
                  "examples": {
                     "application/json": {
                        "status": "UP"
                     }
                  }
               }
            }
         }
      },
      "/translator/translated": {
         "get": {
            "operationId": "getTranslated",
            "description": "traduz uma mensagem",
            "parameters": [
               {
                 "name": "text",
                 "in": "query",
                 "description": "text",
                 "required": true,
                 "type": "string",
                 "x-example": "How are you?"
               },
               {
                 "name": "de",
                 "in": "query",
                 "description": "de",
                 "required": true,
                 "type": "string",
                 "x-example": "en"
               },
               {
                 "name": "para",
                 "in": "query",
                 "description": "para",
                 "required": true,
                 "type": "string",
                 "x-example": "pt"
               }
             ],"responses": {
               "200": {
                  "description": "Text traduzido",
                  "schema": {
                     "$ref": "#/definitions/tranlated"
                  },
                  "examples": {
                     "application/json": {
                        "text": "Como você está?"
                     }
                  }
               }
            }
         }
      }
   },
   "definitions": {
      "healthResponse": {
         "type": "object",
         "properties": {
            "status": {
               "type": "string"
            }
         }
      },
      "tranlated": {
         "type": "object",
         "properties": {
            "status": {
               "type": "string"
            }
         }
      }
   }
}